style(table): remove deprecated `VerticalTable`
