--- client/app/lib/components/core/table/index.tsx
@@ -1,3 +1,0 @@
-export { default as TableContainer } from './TableContainer';
-export type { ColumnTemplate } from './template';
-export { default as VerticalTable } from './VerticalTable';
